refactor : migrate to material theming
This commit addresses issues pointed out on Pr 92.
Also it makes use of a consistent color scheme across the app for
both the light and dark variants of the app and defines material color
attributes.
Small tweaks have been made to have consistent theming such as:
1. SwitchCompat over Switch to have the switch track color blend in
without need for transparency.
2. Removal of duplicate resources and rather changing the fill color
of resource depending on light or dark theme.
3. On the feedback input field,use of a transparent background so as
to show the hint in both light and dark mode,was initially hidden on dark
mode

Fixes #89